fixed input type error in main

+++ __main__.py
@@ -7,7 +7,7 @@ if __name__ == "__main__":
     crp.crop_all()
 
     proportion = 0.3 # Default value
-    proportion = int(input("Proportion of test images (float or int) : "))
+    proportion = float(input("Proportion of test images (float or int) : "))
 
     split.split(('true','false'), proportion)
